+++ BluenetLib/containerClasses/BluenetSettings.swift
@@ -34,9 +34,9 @@ public struct DevicePreferences {
     public var ignoreForBehaviour: Bool = false
     public var useBackgroundBroadcasts: Bool = false
     public var useBaseBroadcasts: Bool = false
-    public var trackingNumber: UInt64 = 0
+    public var trackingNumber: UInt32 = 0
     
-    public init(rssiOffset:Int8? = nil, tapToToggle:Bool? = nil, ignoreForBehaviour:Bool? = nil, useBackgroundBroadcasts:Bool? = nil, useBaseBroadcasts:Bool? = nil, trackingNumber: UInt64? = nil) {
+    public init(rssiOffset:Int8? = nil, tapToToggle:Bool? = nil, ignoreForBehaviour:Bool? = nil, useBackgroundBroadcasts:Bool? = nil, useBaseBroadcasts:Bool? = nil, trackingNumber: UInt32? = nil) {
         if let rssiOffsetValue  = rssiOffset                            { self.rssiOffset              = rssiOffsetValue  }
         if let tapToToggleValue = tapToToggle                           { self.tapToToggle             = tapToToggleValue }
         if let ignoreForBehaviourValue = ignoreForBehaviour             { self.ignoreForBehaviour      = ignoreForBehaviourValue }
@@ -90,7 +90,7 @@ public class BluenetSettings {
         self.locationState.referenceId = referenceId
     }
     
-    public func setDevicePreferences(rssiOffset: Int8, tapToToggle: Bool, ignoreForBehaviour: Bool, useBackgroundBroadcasts: Bool, useBaseBroadcasts: Bool, trackingNumber: UInt64) {
+    public func setDevicePreferences(rssiOffset: Int8, tapToToggle: Bool, ignoreForBehaviour: Bool, useBackgroundBroadcasts: Bool, useBaseBroadcasts: Bool, trackingNumber: UInt32) {
         self.devicePreferences.rssiOffset = rssiOffset
         self.devicePreferences.tapToToggle = tapToToggle
         self.devicePreferences.ignoreForBehaviour = ignoreForBehaviour
